+++ sdk/go/alicloud/config/pulumiTypes.go
@@ -81,6 +81,7 @@ func (o AssumeRoleOutput) SessionName() pulumi.StringPtrOutput {
 type Endpoints struct {
 	Actiontrail     *string `pulumi:"actiontrail"`
 	Adb             *string `pulumi:"adb"`
+	Alidns          *string `pulumi:"alidns"`
 	Alikafka        *string `pulumi:"alikafka"`
 	Apigateway      *string `pulumi:"apigateway"`
 	Bssopenapi      *string `pulumi:"bssopenapi"`
@@ -141,6 +142,7 @@ type EndpointsInput interface {
 type EndpointsArgs struct {
 	Actiontrail     pulumi.StringPtrInput `pulumi:"actiontrail"`
 	Adb             pulumi.StringPtrInput `pulumi:"adb"`
+	Alidns          pulumi.StringPtrInput `pulumi:"alidns"`
 	Alikafka        pulumi.StringPtrInput `pulumi:"alikafka"`
 	Apigateway      pulumi.StringPtrInput `pulumi:"apigateway"`
 	Bssopenapi      pulumi.StringPtrInput `pulumi:"bssopenapi"`
@@ -244,6 +246,10 @@ func (o EndpointsOutput) Actiontrail() pulumi.StringPtrOutput {
 
 func (o EndpointsOutput) Adb() pulumi.StringPtrOutput {
 	return o.ApplyT(func(v Endpoints) *string { return v.Adb }).(pulumi.StringPtrOutput)
+}
+
+func (o EndpointsOutput) Alidns() pulumi.StringPtrOutput {
+	return o.ApplyT(func(v Endpoints) *string { return v.Alidns }).(pulumi.StringPtrOutput)
 }
 
 func (o EndpointsOutput) Alikafka() pulumi.StringPtrOutput {