+++ sdk/go/alicloud/pulumiTypes.go
@@ -456,6 +456,7 @@ func (o ProviderAssumeRolePtrOutput) SessionName() pulumi.StringPtrOutput {
 type ProviderEndpoint struct {
 	Actiontrail     *string `pulumi:"actiontrail"`
 	Adb             *string `pulumi:"adb"`
+	Alidns          *string `pulumi:"alidns"`
 	Alikafka        *string `pulumi:"alikafka"`
 	Apigateway      *string `pulumi:"apigateway"`
 	Bssopenapi      *string `pulumi:"bssopenapi"`
@@ -516,6 +517,7 @@ type ProviderEndpointInput interface {
 type ProviderEndpointArgs struct {
 	Actiontrail     pulumi.StringPtrInput `pulumi:"actiontrail"`
 	Adb             pulumi.StringPtrInput `pulumi:"adb"`
+	Alidns          pulumi.StringPtrInput `pulumi:"alidns"`
 	Alikafka        pulumi.StringPtrInput `pulumi:"alikafka"`
 	Apigateway      pulumi.StringPtrInput `pulumi:"apigateway"`
 	Bssopenapi      pulumi.StringPtrInput `pulumi:"bssopenapi"`
@@ -619,6 +621,10 @@ func (o ProviderEndpointOutput) Actiontrail() pulumi.StringPtrOutput {
 
 func (o ProviderEndpointOutput) Adb() pulumi.StringPtrOutput {
 	return o.ApplyT(func(v ProviderEndpoint) *string { return v.Adb }).(pulumi.StringPtrOutput)
+}
+
+func (o ProviderEndpointOutput) Alidns() pulumi.StringPtrOutput {
+	return o.ApplyT(func(v ProviderEndpoint) *string { return v.Alidns }).(pulumi.StringPtrOutput)
 }
 
 func (o ProviderEndpointOutput) Alikafka() pulumi.StringPtrOutput {