+++ sdk/nodejs/resourcemanager/index.ts
@@ -4,7 +4,11 @@
 // Export members:
 export * from "./account";
 export * from "./folder";
+export * from "./getFolders";
+export * from "./getResourceGroups";
 export * from "./handshake";
 export * from "./policy";
+export * from "./policyVersion";
+export * from "./resourceDirectory";
 export * from "./resourceGroup";
 export * from "./role";