+++ sdk/python/pulumi_alicloud/dns/domain_group.py
@@ -9,17 +9,38 @@ import pulumi.runtime
 from typing import Union
 from .. import utilities, tables
 
-class Group(pulumi.CustomResource):
-    name: pulumi.Output[str]
+class DomainGroup(pulumi.CustomResource):
+    group_name: pulumi.Output[str]
     """
-    Name of the domain group.    
+    Name of the domain group. 
     """
-    def __init__(__self__, resource_name, opts=None, name=None, __props__=None, __name__=None, __opts__=None):
+    lang: pulumi.Output[str]
+    """
+    User language. 
+    """
+    def __init__(__self__, resource_name, opts=None, group_name=None, lang=None, __props__=None, __name__=None, __opts__=None):
         """
-        Create a Group resource with the given unique name, props, and options.
+        Provides a Alidns Domain Group resource. For information about Alidns Domain Group and how to use it, see [What is Resource Alidns Domain Group](https://www.alibabacloud.com/help/en/doc-detail/29762.htm).
+
+        > **NOTE:** Available in v1.84.0+.
+
+        ## Example Usage
+
+
+
+        ```python
+        import pulumi
+        import pulumi_alicloud as alicloud
+
+        # Add a new Alinds Domain Group.
+        example = alicloud.dns.DomainGroup("example", group_name="tf-testDG")
+        ```
+
+
         :param str resource_name: The name of the resource.
         :param pulumi.ResourceOptions opts: Options for the resource.
-        :param pulumi.Input[str] name: Name of the domain group.    
+        :param pulumi.Input[str] group_name: Name of the domain group. 
+        :param pulumi.Input[str] lang: User language. 
         """
         if __name__ is not None:
             warnings.warn("explicit use of __name__ is deprecated", DeprecationWarning)
@@ -38,30 +59,35 @@ class Group(pulumi.CustomResource):
                 raise TypeError('__props__ is only valid when passed in combination with a valid opts.id to get an existing resource')
             __props__ = dict()
 
-            __props__['name'] = name
-        super(Group, __self__).__init__(
-            'alicloud:dns/group:Group',
+            if group_name is None:
+                raise TypeError("Missing required property 'group_name'")
+            __props__['group_name'] = group_name
+            __props__['lang'] = lang
+        super(DomainGroup, __self__).__init__(
+            'alicloud:dns/domainGroup:DomainGroup',
             resource_name,
             __props__,
             opts)
 
     @staticmethod
-    def get(resource_name, id, opts=None, name=None):
+    def get(resource_name, id, opts=None, group_name=None, lang=None):
         """
-        Get an existing Group resource's state with the given name, id, and optional extra
+        Get an existing DomainGroup resource's state with the given name, id, and optional extra
         properties used to qualify the lookup.
 
         :param str resource_name: The unique name of the resulting resource.
         :param str id: The unique provider ID of the resource to lookup.
         :param pulumi.ResourceOptions opts: Options for the resource.
-        :param pulumi.Input[str] name: Name of the domain group.    
+        :param pulumi.Input[str] group_name: Name of the domain group. 
+        :param pulumi.Input[str] lang: User language. 
         """
         opts = pulumi.ResourceOptions.merge(opts, pulumi.ResourceOptions(id=id))
 
         __props__ = dict()
 
-        __props__["name"] = name
-        return Group(resource_name, opts=opts, __props__=__props__)
+        __props__["group_name"] = group_name
+        __props__["lang"] = lang
+        return DomainGroup(resource_name, opts=opts, __props__=__props__)
     def translate_output_property(self, prop):
         return tables._CAMEL_TO_SNAKE_CASE_TABLE.get(prop) or prop